Extract APIBasedTestFileCompilerBase
