+++ src/compiler/index.js
@@ -2,6 +2,7 @@ import * as fs from 'fs';
 import Promise from 'pinkie';
 import { flattenDeep as flatten, find, chunk } from 'lodash';
 import stripBom from 'strip-bom';
+import sourceMapSupport from 'source-map-support';
 import { Compiler as LegacyTestFileCompiler } from 'testcafe-legacy-api';
 import hammerhead from 'testcafe-hammerhead';
 import EsNextTestFileCompiler from './test-file/formats/es-next';
@@ -23,6 +24,16 @@ export default class Compiler {
             new EsNextTestFileCompiler(),
             new RawTestFileCompiler()
         ];
+
+        Compiler._setupSourceMapsSupport();
+    }
+
+    static _setupSourceMapsSupport () {
+        sourceMapSupport.install({
+            hookRequire:              true,
+            handleUncaughtExceptions: false,
+            environment:              'node'
+        });
     }
 
     async _compileTestFile (filename) {
