+++ src/compiler/test-file/base.js
