+++ src/compiler/test-file/formats/es-next.js
@@ -1,41 +1,10 @@
-import { dirname, join, relative, sep as pathSep } from 'path';
-import { readFileSync } from 'fs';
-import stripBom from 'strip-bom';
-import sourceMapSupport from 'source-map-support';
 import loadBabelLibs from '../../load-babel-libs';
-import TestFileCompilerBase from '../compiler-base';
-import TestFile from '../../../api/structure/test-file';
-import Fixture from '../../../api/structure/fixture';
-import Test from '../../../api/structure/test';
-import { TestCompilationError, APIError } from '../../../errors/runtime';
-import stackCleaningHook from '../../../errors/stack-cleaning-hook';
+import APIBasedTestFileCompilerBase from '../api-based';
 
-const EXPORTABLE_LIB_PATH = join(__dirname, '../../../api/exportable-lib');
-const CWD                 = process.cwd();
-
-const FIXTURE_RE       = /(^|;|\s+)fixture\s*(\.|\(|`)/;
-const TEST_RE          = /(^|;|\s+)test\s*(\.|\()/;
 const BABEL_RUNTIME_RE = /^babel-runtime(\\|\/|$)/;
 
-var Module = module.constructor;
-
-export default class ESNextTestFileCompiler extends TestFileCompilerBase {
-    constructor () {
-        super();
-
-        this.sourceMaps = {};
-        this.cache      = {};
-
-        this._setupSourceMapsSupport();
-    }
-
-    static _getNodeModulesLookupPath (filename) {
-        var dir = dirname(filename);
-
-        return Module._nodeModulePaths(dir);
-    }
-
-    static _getBabelOptions (filename) {
+export default class ESNextTestFileCompiler extends APIBasedTestFileCompilerBase {
+    _getBabelOptions (filename) {
         var { presetStage2, transformRuntime, presetEnv } = loadBabelLibs();
 
         // NOTE: passPrePreset and complex presets is a workaround for https://github.com/babel/babel/issues/2877
@@ -57,7 +26,7 @@ export default class ESNextTestFileCompiler extends TestFileCompilerBase {
 
             resolveModuleSource: source => {
                 if (source === 'testcafe')
-                    return EXPORTABLE_LIB_PATH;
+                    return this.EXPORTABLE_LIB_PATH;
 
                 if (BABEL_RUNTIME_RE.test(source)) {
                     try {
@@ -73,143 +42,25 @@ export default class ESNextTestFileCompiler extends TestFileCompilerBase {
         };
     }
 
-    static _isNodeModulesDep (filename) {
-        return relative(CWD, filename)
-                   .split(pathSep)
-                   .indexOf('node_modules') >= 0;
-    }
-
-    static _execAsModule (code, filename) {
-        var mod = new Module(filename, module.parent);
-
-        mod.filename = filename;
-        mod.paths    = ESNextTestFileCompiler._getNodeModulesLookupPath(filename);
-
-        mod._compile(code, filename);
-    }
-
-    _setupSourceMapsSupport () {
-        sourceMapSupport.install({
-            hookRequire:              true,
-            handleUncaughtExceptions: false,
-            environment:              'node'
-        });
-    }
-
-    _compileES (code, filename) {
+    _compileCode (code, filename) {
         var { babel } = loadBabelLibs();
 
         if (this.cache[filename])
             return this.cache[filename];
 
-        var opts     = ESNextTestFileCompiler._getBabelOptions(filename);
+        var opts     = this._getBabelOptions(filename);
         var compiled = babel.transform(code, opts);
 
-        this.cache[filename]      = compiled.code;
-        this.sourceMaps[filename] = compiled.map;
+        this.cache[filename] = compiled.code;
 
         return compiled.code;
     }
 
-    _setupRequireHook (testFile) {
-        var origRequireExtension = require.extensions['.js'];
-
-        require.extensions['.js'] = (mod, filename) => {
-            // NOTE: remove global API so that it will be unavailable for the dependencies
-            this._removeGlobalAPI();
-
-            if (ESNextTestFileCompiler._isNodeModulesDep(filename))
-                origRequireExtension(mod, filename);
-            else {
-                var code         = readFileSync(filename);
-                var compiledCode = this._compileES(stripBom(code), filename);
-
-                mod.paths = ESNextTestFileCompiler._getNodeModulesLookupPath(filename);
-
-                mod._compile(compiledCode, filename);
-            }
-
-            this._addGlobalAPI(testFile);
-        };
-
-        return origRequireExtension;
-    }
-
-    _compileESForTestFile (code, filename) {
-        var compiledCode = null;
-
-        stackCleaningHook.enabled = true;
-
-        try {
-            compiledCode = this._compileES(code, filename);
-        }
-        catch (err) {
-            throw new TestCompilationError(err);
-        }
-        finally {
-            stackCleaningHook.enabled = false;
-        }
-
-        return compiledCode;
-    }
-
-    _addGlobalAPI (testFile) {
-        Object.defineProperty(global, 'fixture', {
-            get:          () => new Fixture(testFile),
-            configurable: true
-        });
-
-        Object.defineProperty(global, 'test', {
-            get:          () => new Test(testFile),
-            configurable: true
-        });
-    }
-
-    _removeGlobalAPI () {
-        delete global.fixture;
-        delete global.test;
-    }
-
-    _hasTests (code) {
-        return FIXTURE_RE.test(code) && TEST_RE.test(code);
+    _getRequireCompilers () {
+        return { '.js': (code, filename) => this._compileCode(code, filename) };
     }
 
     getSupportedExtension () {
         return '.js';
-    }
-
-    compile (code, filename) {
-        var compiledCode = this._compileESForTestFile(code, filename);
-        var testFile     = new TestFile(filename);
-
-        this._addGlobalAPI(testFile);
-
-        stackCleaningHook.enabled = true;
-
-        var origRequireExtension = this._setupRequireHook(testFile);
-
-        try {
-            ESNextTestFileCompiler._execAsModule(compiledCode, filename);
-        }
-        catch (err) {
-            // HACK: workaround for the `instanceof` problem
-            // (see: http://stackoverflow.com/questions/33870684/why-doesnt-instanceof-work-on-instances-of-error-subclasses-under-babel-node)
-            if (err.constructor !== APIError)
-                throw new TestCompilationError(err);
-
-            throw err;
-        }
-        finally {
-            require.extensions['.js'] = origRequireExtension;
-            stackCleaningHook.enabled = false;
-
-            this._removeGlobalAPI();
-        }
-
-        return testFile.getTests();
-    }
-
-    cleanUpCache () {
-        this.cache = null;
     }
 }
